refactor: centralize value conversion utility
