+++ src/tnfr/collections_utils.py
@@ -17,6 +17,7 @@ import logging
 import math
 from itertools import islice
 
+from .value_utils import _convert_value
 T = TypeVar("T")
 
 logger = logging.getLogger(__name__)
@@ -56,30 +57,6 @@ def ensure_collection(
         return data
     except TypeError as exc:
         raise TypeError(f"{it!r} is not iterable") from exc
-
-
-def _convert_value(
-    value: Any,
-    conv: Callable[[Any], T],
-    *,
-    strict: bool = False,
-    key: str | None = None,
-    log_level: int | None = None,
-) -> tuple[bool, T | None]:
-    """Try to convert ``value`` using ``conv`` handling errors."""
-    try:
-        return True, conv(value)
-    except (ValueError, TypeError) as exc:
-        level = log_level if log_level is not None else (
-            logging.ERROR if strict else logging.DEBUG
-        )
-        if key is not None:
-            logger.log(level, "No se pudo convertir el valor para %r: %s", key, exc)
-        else:
-            logger.log(level, "No se pudo convertir el valor: %s", exc)
-        if strict:
-            raise
-        return False, None
 
 
 def normalize_weights(
