+++ src/tnfr/helpers.py
@@ -49,6 +49,7 @@ from .collections_utils import (
     normalize_counter,
     mix_groups,
 )
+from .value_utils import _convert_value
 
 T = TypeVar("T")
 
@@ -176,37 +177,6 @@ def angle_diff(a: float, b: float) -> float:
 # -------------------------
 # Acceso a atributos con alias
 # -------------------------
-
-
-def _convert_value(
-    value: Any,
-    conv: Callable[[Any], T],
-    *,
-    strict: bool = False,
-    key: str | None = None,
-    log_level: int | None = None,
-) -> tuple[bool, T | None]:
-    """Intenta convertir ``value`` usando ``conv`` manejando errores.
-
-    ``log_level`` controla el nivel de logging cuando la conversión falla en
-    modo laxo. Por defecto se usa ``logging.ERROR`` si ``strict`` es ``True`` y
-    ``logging.DEBUG`` en caso contrario.
-    """
-    try:
-        return True, conv(value)
-    except (ValueError, TypeError) as exc:
-        level = log_level if log_level is not None else (
-            logging.ERROR if strict else logging.DEBUG
-        )
-        if key is not None:
-            logger.log(level, "No se pudo convertir el valor para %r: %s", key, exc)
-        else:
-            logger.log(level, "No se pudo convertir el valor: %s", exc)
-        if strict:
-            raise
-        return False, None
-
-
 def _validate_aliases(aliases: Sequence[str]) -> Sequence[str]:
     """Return ``aliases`` ensuring it's a non-empty sequence of strings."""
